refactor(console): extract the approval card's body for reuse (#379)
Two surfaces are about to decide the same approval — the page and the
conversation it was raised in — and they must say the same thing. Two
copies of "what will happen, who is asking, what it is for, how long it has
waited" would drift, and the half that drifted would be the one nobody was
reading when it mattered.

The action buttons stay with their owners: the page resolves with the
default response shape, the inline card detaches. Same content, different
verbs. `ApprovalsView` renders identically to before.
